ClipExecutor is renamed
